more size_t/int twiddling

+++ src/utils/pathut.h
@@ -117,7 +117,7 @@ bool path_readable(const std::string& path);
 
 bool wchartoutf8(const wchar_t *in, std::string& out, int len = 0);
 std::string wchartoutf8(const wchar_t *in, int len = 0);
-bool utf8towchar(const std::string& in, wchar_t *out, size_t obytescap);
+bool utf8towchar(const std::string& in, wchar_t *out, int obytescap);
 std::unique_ptr<wchar_t[]> utf8towchar(const std::string& in);
 
 // Convert between slash and backslash separators.
